requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alpha-vantage==2.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
beautifulsoup4==4.12.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.28.0
dvc-data==2.20.0
dvc-gdrive==2.20.0
dvc-http==2.30.2
dvc-objects==1.1.0
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.46.0
frozendict==2.4.4
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.61.0
grandalf==0.8
gto==1.5.0
html5lib==1.1
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
iterative-telemetry==0.0.8
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.3
llvmlite==0.43.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
networkx==3.2.1
numba==0.60.0
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.5
orjson==3.9.10
packaging==23.2
pandas==2.2.1
pandas-datareader==0.10.0
pathfinding==1.0.4
pathspec==0.11.2
peewee==3.17.6
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydot==1.4.2
PyDrive2==1.17.0
pygit2==1.13.2
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.6.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.3
scmrepo==1.4.1
seaborn==0.13.2
semver==3.0.2
shap==0.46.0
shortuuid==1.0.11
shtab==1.6.4
sinch==0.3.1
six==1.16.0
slicer==0.0.8
smmap==5.0.1
soupsieve==2.5
sqltrie==0.8.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tomlkit==0.12.2
tqdm==4.66.1
twilio==8.12.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
voluptuous==0.13.1
vonage==3.13.0
vonage-jwt==1.1.0
wcwidth==0.2.9
webencodings==0.5.1
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.2
yfinance==0.2.43
zc.lockfile==3.0.post1
